Move KGA handler to separate file
