+++ src/scripts/main.js
@@ -8,6 +8,7 @@ import * as detailers from './modules/details';
 import * as stagers from './modules/stages';
 import * as stagingFuncs from './modules/stages/legacy';
 import * as detailingFuncs from './modules/details/legacy';
+import {recordPrizePack} from "./modules/ajax-handlers/legacy";
 
 (function () {
     'use strict';
@@ -351,7 +352,7 @@ import * as detailingFuncs from './modules/details/legacy';
             } else if (url.includes("mousehuntgame.com/managers/ajax/events/kings_giveaway.php")) {
                 // Triggers on Birthday Items claim, room change click (+others, perhaps).
                 // Wed Jun 23 2021 22:00:00 GMT-0400 [King's Giveaway Key Vanishing date 15th])
-                getSettings(settings => recordPrizePack(settings, xhr));
+                recordPrizePack(xhr, logger, submitConvertible);
             } else if (url.includes("mousehuntgame.com/managers/ajax/users/session.php")) {
                 createHunterIdHash();
             }
@@ -428,58 +429,6 @@ import * as detailingFuncs from './modules/details/legacy';
             "crowns": payload,
             "settings": settings,
         }, window.origin);
-    }
-
-    /**
-     * Record Mini Prize Pack convertible submissions as convertibles in MHCT
-     * @param {Object <string, any>} settings The user's extension settings.
-     * @param {JQuery.jqXHR} xhr jQuery-wrapped XMLHttpRequest object encapsulating the http request to the remote server (HG).
-     */
-    function recordPrizePack(settings, xhr) {
-        if (
-            !xhr.responseJSON?.kings_giveaway_result ||
-            !xhr.responseJSON.inventory || !xhr.responseJSON.kings_giveaway_result.quantity ||
-            xhr.responseJSON.kings_giveaway_result.slot !== "bonus"
-        ) {
-            logger.debug('Skipped mini prize pack submission due to unhandled XHR structure. This is probably fine.');
-            window.postMessage({
-                "mhct_log_request": 1,
-                "is_error": true,
-                "kga_2021_response": xhr.responseJSON,
-                "reason": "Unable to parse kga 2021 response. This is normal if a pack wasn't opened",
-            }, window.origin);
-            return;
-        }
-        const result = xhr.responseJSON.kings_giveaway_result;
-        const inventory = xhr.responseJSON.inventory;
-
-        const convertible = {
-            name: "King's Mini Prize Pack",
-            id: 130008,
-            quantity: result.quantity,
-        };
-
-        const item_map = {
-            "gold_stat_item": 431,
-        };
-
-        const items = [];
-        result.items.forEach(item => {
-            const i = {
-                "type": item.type,
-                "quantity": parseInt((item.quantity + ",").replace(/,/g, ""), 10),
-            };
-            if (item.type in inventory && inventory[item.type].item_id) {
-                i.id = inventory[item.type].item_id;
-            }
-            else if (item.type in item_map) {
-                i.id = item_map[item.type];
-            }
-            items.push(i);
-        });
-
-        logger.debug("Prizepack: ", {convertible, items, settings});
-        submitConvertible(convertible, items);
     }
 
     // Record map mice
